+++ app/controllers/TracingStoreController.scala
@@ -2,11 +2,9 @@ package controllers
 
 import com.mohiva.play.silhouette.api.Silhouette
 import com.scalableminds.util.accesscontext.GlobalAccessContext
-import com.scalableminds.util.tools.{Fox, FoxImplicits}
+import com.scalableminds.util.tools.FoxImplicits
 import javax.inject.Inject
 import models.annotation.{TracingStore, TracingStoreDAO, TracingStoreService}
-import models.binary.{DataStore, DataStoreDAO, DataStoreService}
-import models.user.UserService
 import net.liftweb.common.Empty
 import oxalis.security.WkEnv
 import play.api.i18n.Messages