+++ conf/evolutions/047-add-tracingstore-publicUrl.sql
@@ -1,4 +1,4 @@
--- https://github.com/scalableminds/webknossos/pull/X
+-- https://github.com/scalableminds/webknossos/pull/4286
 
 START TRANSACTION;
 
@@ -12,5 +12,7 @@ UPDATE webknossos.tracingStores SET publicUrl = url;
 
 ALTER TABLE webknossos.tracingStores ALTER COLUMN publicUrl SET NOT NULL;
 ALTER TABLE webknossos.tracingStores ADD CHECK (publicUrl ~* '^https?://[a-z0-9\.]+.*$');
+
+UPDATE webknossos.releaseInformation SET schemaVersion = 47;
 
 COMMIT TRANSACTION;